+++ test/common.dart
@@ -1,0 +1,3 @@
+import 'package:test/test.dart';
+
+myTest(String name, Function testFn) => test(name, testFn);